+++ backends/cuda/ceed-cuda-common.c
@@ -9,6 +9,23 @@
 #include "ceed-cuda-common.h"
 
 //------------------------------------------------------------------------------
+// Get root resource without device spec
+//------------------------------------------------------------------------------
+int CeedCudaGetResourceRoot(Ceed ceed, const char *resource,
+                            char **resource_root) {
+  int ierr;
+
+  char *device_spec = strstr(resource, ":device_id=");
+  size_t resource_root_len = device_spec
+                             ? (size_t)(device_spec - resource) + 1
+                             : strlen(resource) + 1;
+  ierr = CeedCalloc(resource_root_len, resource_root); CeedChkBackend(ierr);
+  memcpy(*resource_root, resource, resource_root_len - 1);
+
+  return CEED_ERROR_SUCCESS;
+}
+
+//------------------------------------------------------------------------------
 // Device information backend init
 //------------------------------------------------------------------------------
 int CeedCudaInit(Ceed ceed, const char *resource) {